requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-telegram-bot[webhooks]==21.1.1
sqlalchemy==2.0.30
asyncpg==0.29.0